+++ astropy/units/tests/test_equivalencies.py
@@ -515,3 +515,10 @@ def test_temperature():
     t_k = 0 * u.K
     assert_allclose(t_k.to(u.deg_C, u.temperature()).value, -273.15)
     assert_allclose(t_k.to(deg_F, u.temperature()).value, -459.67)
+
+def test_temperature_energy():
+    from ... import constants
+    x = 1000 * u.K
+    y = (x * constants.k_B).to(u.keV)
+    assert_allclose(x.to(u.keV, u.temperature_energy()).value, y)           
+    assert_allclose(y.to(u.K, u.temperature_energy()).value, x)