+++ CHANGES.rst
@@ -206,6 +206,9 @@ New Features
   - Added ``temperature()`` equivalencies to support conversion between
     Kelvin, Celsius, and Fahrenheit. [#2209]
 
+  - Added ``temperature_energy()`` equivalencies to support conversion
+    between electron-volt and Kelvin. [#2637]
+
   - The runtime of ``astropy.units.Unit.compose`` is greatly improved
     (by a factor of 2 in most cases) [#2544]